refactor:코드 분리 리팩토링(#21)
