+++ src/apis/utils/navigationErrors.js
@@ -1,0 +1,3 @@
+export function isNavigationNotFoundCode(code) {
+  return typeof code === 'string' && code.startsWith('NAVIGATION404_')
+}
